Support removing crime modifiers and add learning of suspects.
It is now possible to remove modifiers for crimes by clicking on the
modifier button again.

Additionally, the application has now gained the ability to learn the
names and phone numbers of suspects. Names and phone numbers are learned
whenever you generate a report.

You can manipulate the list of suspects and phone numbers yourself while
the program is closed by editing Suspects.txt